Refactor the StartAvaloniaPreviewerAction, add file system watcher (#6, #8)
